refactor: 사용자 정보 추출 securityContext 에서 하도록 수정
