--- mokakbob-api/src/main/java/com/mokakbob/global/support/AuthConstants.java
@@ -1,5 +1,0 @@
-package com.mokakbob.global.support;
-
-public class AuthConstants {
-    public static final String TOKEN_ATTRIBUTE = "memberId";
-}
